Updated orders to have amount as a float and not a int

+++ api/models/orders.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, DATETIME
+from sqlalchemy import Column, ForeignKey, Integer, String, DATETIME, DECIMAL
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from ..dependencies.database import Base
@@ -12,7 +12,7 @@ class Order(Base):
     order_date = Column(DATETIME, nullable=False, server_default=str(datetime.now()))
     description = Column(String(300))
     sandwich_id = Column(Integer, ForeignKey("sandwiches.id"), nullable=False)
-    amount = Column(Integer, index=True, nullable=False)
+    amount = Column(DECIMAL, index=True, nullable=False)
     restaurant_id = Column(Integer, ForeignKey("restaurants.id"), nullable=False)
     delivery_method = Column(String(300), nullable=False)
     status_of_order = Column(String(300), nullable=False, server_default="pending")
